+++ src/events/Interactions/commandCreate.ts
@@ -6,13 +6,12 @@ export default createEvent<"interactionCreate">({
     once: false
 }, async (interaction) => {
     if (!interaction.isChatInputCommand()) return;
-    if (!interaction.inCachedGuild()) return;
-    
-    const { client, user, options } = interaction;
+    if (!interaction.inCachedGuild()) return;  
 
+    const { client, user, options } = interaction;
     const command = client.commands.get(interaction.commandName);
     if (!command) return;
-
+    
     let guild = await prisma
     .guild
     .findUnique({ where: { guildId: interaction.guildId }, include: { settings: true } });
@@ -26,19 +25,19 @@ export default createEvent<"interactionCreate">({
     client.logger.info('event(interactionCreate)', `${user.username} used /${interaction.commandName}`);
 
     try {
-        let command_path;
-        if (options.getSubcommand()) {
-            command_path = command.options?.path?.find(kv => kv.K === options.getSubcommand())?.V;
-        } else {
-            command_path = command.options?.path?.find(kv => kv.K === interaction.commandName)?.V;
-        }
-
         const context: CommandContext = {
+            client,
             interaction,
             prisma: prisma
         };
 
-        (await import(`../../commands/sub/${command_path}`)).run(context);
+        if (options.getSubcommand(false)) {
+            const path = command.options?.path?.find(p => p.K === options.getSubcommand())?.V;
+            (await import(`../../commands/sub/${path}`)).run(context);
+        } else {
+            const path = command.options?.path?.find(p => p.K === interaction.commandName)?.V;
+            (await import(`../../commands/${path}`)).default.run(context);
+        }
     } catch (error) {
         client.logger.error("event(interactionCreate)", error);
     }