+++ INFO/changelog.md
@@ -1,3 +1,27 @@
+# @0.2.5 (& @0.2.7)
+> *✨ | Update (0.2.5)*
+
+- **Developed By:** @0sapphy
+- **Release Date:** 25/11/2024
+
+## New
+
+- Database (using prisma/client)
+- Commands [/settings welcome|/ping]
+- Example .env 
+
+## Updated
+
+- Command event (finds command files differently.)
+- `createCommand()` `run` param, accepts `CommandContext`.
+- Added client to `CommandContext`.
+- (prisma): Added logging [query|info|warn|error]
+
+## Removed
+
+- Command checking (`registerApplicationCommands()`)
+- Import alias
+
 # @0.1.12
 > *The Base*